python-service/requirements.txt

flask
elasticsearch
google-generativeai
python-dotenv
